requirements.txt
# ── Warehouse & transformation ────────────────────────────────────────────────
duckdb>=0.10.0
dbt-core>=1.7.0
dbt-duckdb>=1.7.0

# ── Data manipulation ─────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0

# ── Machine learning ──────────────────────────────────────────────────────────
scikit-learn>=1.3.0
xgboost>=2.0.0
lifetimes>=0.11.3
joblib>=1.3.0

# ── Explainability ────────────────────────────────────────────────────────────
shap>=0.44.0
matplotlib>=3.7.0  # SHAP plots only — do NOT use in Streamlit app

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit>=1.30.0
plotly>=5.18.0

# ── Data quality ──────────────────────────────────────────────────────────────
great-expectations>=0.18.0

# ── Testing & dev ─────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0

# ── Notebooks ─────────────────────────────────────────────────────────────────
ipykernel>=6.0.0
nbformat>=4.2.0
